python/flask-sqlite/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#    pipenv lock --requirements
#

-i https://pypi.org/simple
flask==1.1.2
SQLAlchemy~=1.3.22
python-dotenv~=0.15.0
Werkzeug~=1.0.1
alembic~=0.9.6